# [pt][quant] qmul and qadd should preserve input memory format